Loads of fixes & beta testing. Version now 3.0.0.5 Fixed settings menu not disabling. Fixed update checks. Fixed First run updates. Fixed a bug with auto finding installations (null reg key) related to non-steam installations. These are now found properly. Moved the help link to a button to be easier to stop. Made the help open automatically if theres no installations. Added more logging around potential issues with auto-detection. Added a catch for steam finding failure and to catch an issue where old steam installations were pointing to wrong file. Added innosetup script for creating the installer. Added .Net Core 2.0 project for building the hash and compressing the files from on the servers. Added ability to open the "news" in the browser. Various other bug fixes I cant remember.

+++ updater-lib/UpdateHash.cs
@@ -32,10 +32,10 @@ namespace cs_updater_lib
             Source = source;
         }
 
-        public int getFileCount()
+        public double getFileCount()
         {
             if (this.Files == null) return 0;
-            int count = this.Files.Count();
+            double count = this.Files.Count();
             foreach (UpdateHashItem i in this.Files)
             {
                 count += i.getFilesCount();
